requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
evdev==1.9.2
flatbuffers==25.9.23
fonttools==4.61.0
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.4
MouseInfo==0.1.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
PyAutoGUI==0.9.54
pycparser==2.23
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pynput==1.8.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
